lab1.2/requirements.txt
Django==4.2.4
django-minio-storage==0.5.6
djangorestframework==3.14.0